@@ -7,12 +7,13 @@
     "test": "jest"
   },
   "dependencies": {
-    "react": "15.3.2",
-    "react-native": "^0.37.0",
+    "react": "15.4.1",
+    "react-native": "^0.38.0",
     "react-native-db-models": "^0.1.3",
     "react-native-side-menu": "^0.20.1",
     "react-redux": "^4.4.6",
-    "redux": "^3.6.0"
+    "redux": "^3.6.0",
+    "tcomb-form-native": "^0.6.1"
   },
   "jest": {
     "preset": "jest-react-native"
